STUD-28 - Добавлен institutes

+++ src/main/java/ru/urfu/sv/studentvoice/model/repository/InstituteRepository.java
@@ -7,11 +7,13 @@ import ru.urfu.sv.studentvoice.model.domain.entity.Institute;
 import java.util.Optional;
 
 @Repository
-public interface InstituteRepository extends JpaRepository<Institute, Integer> {
+public interface InstituteRepository extends JpaRepository<Institute, Long> {
+
     Optional<Institute> findByFullNameIgnoreCase(String fullName);
 
     Optional<Institute> findByShortNameIgnoreCase(String shortName);
 
     boolean existsByAddress(String address);
+
     Optional<Institute> findByAddressIgnoreCase(String address);
 }
